+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fix fatal error on identity page if Enigma plugin is misconfigured (#8719)
 - Fix so N property always exists in a vCard export (#8771)
 - Fix authenticating to Courier IMAP with passwords containing a '~' character (#8772)
+- Fix handling of smtp/imap port options on configuration file update (#8756)
 
 ## Release 1.6.0